requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.44.0
anyio==4.8.0
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.5
Brlapi==0.8.4
bs4==0.0.2
cachetools==5.5.0
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.8
crit==3.17.1
cryptography==38.0.4
cupshelpers==1.0
databases==0.9.0
dbus-python==1.3.2
decorator==5.2.1
defusedxml==0.7.1
distro==1.8.0
distro-info==1.5+deb12u1
dnspython==2.7.0
docopt==0.6.2
durationpy==0.9
ecdsa==0.19.0
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.6
fastjsonschema==2.21.1
frozenlist==1.5.0
google-auth==2.37.0
graphviz==0.20.3
greenlet==3.1.1
gyp==0.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.3
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kubernetes==31.0.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
libvirt-python==9.0.0
louis==3.24.0
Mako==1.2.4.dev0
MarkupSafe==2.1.2
matplotlib-inline==0.1.7
mcp==1.2.0
mistune==3.1.2
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
oauthlib==3.2.2
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==4.21.12
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycups==2.0.1
pycurl==7.45.2
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyGObject==3.42.2
PyJWT==2.6.0
pyparsing==3.0.9
PySimpleSOAP==1.16.2
pysmbc==1.0.23
python-apt==2.6.0
python-dateutil==2.8.2
python-debian==0.1.49
python-debianbts==4.0.1
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
pyxdg==0.28
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
reportbug==12.0.0
requests==2.28.1
requests-oauthlib==2.0.0
rpds-py==0.23.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.41.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==1.26.12
uvicorn==0.34.0
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
xdg==5
yarg==0.1.9
yarl==1.18.3
